+++ circuitbreaker.go
@@ -7,19 +7,23 @@
 //
 // Circuit includes three types of circuit breakers:
 //
-// A ThresholdBreaker will trip when the failure count reaches a given threshold.
-// It does not matter how long it takes to reach the threshold.
+// A Threshold Breaker will trip when the failure count reaches a given threshold.
+// It does not matter how long it takes to reach the threshold and the failures do
+// not need to be consecutive.
 //
-// A FrequencyBreaker will trip when the failure count reaches a given threshold
-// within a given time period.
+// A Consecutive Breaker will trip when the consecutive failure count reaches a given
+// threshold. It does not matter how long it takes to reach the threshold, but the
+// failures do need to be consecutive.
 //
-// A TimeoutBreaker will trip when the failure count reaches a given threshold, with
-// the added feature that the remote call taking longer than a given timeout will
-// count as a failure.
 //
-// Other types of circuit breakers can be easily built. Embedding a TrippableBreaker
-// struct and providing the failure semantics with custom Fail() and Call() functions
-// are all that is typically needed.
+// When wrapping blocks of code with a Breaker's Call() function, a time out can be
+// specified. If the time out is reached, the breaker's Fail() function will be called.
+//
+//
+// Other types of circuit breakers can be easily built by creating a Breaker and
+// adding a custom TripFunc. A TripFunc is called when a Breaker Fail()s and receives
+// the breaker as an argument. It then returns true or false to indicate whether the
+// breaker should trip.
 //
 // The package also provides a wrapper around an http.Client that wraps all of
 // the http.Client functions with a Breaker.
@@ -28,6 +32,7 @@ package circuit
 
 import (
 	"errors"
+	"github.com/cenkalti/backoff"
 	"sync/atomic"
 	"time"
 	"unsafe"
@@ -50,55 +55,89 @@ const (
 	closed   state = iota
 )
 
+var defaultInitialBackOffInterval = 500 * time.Millisecond
+
 // Error codes returned by Call
 var (
 	ErrBreakerOpen    = errors.New("breaker open")
 	ErrBreakerTimeout = errors.New("breaker time out")
 )
 
-var noop = &noOpBreaker{}
+// TripFunc is a function called by a Breaker's Fail() function and determines whether
+// the breaker should trip. It will receive the Breaker as an argument and returns a
+// boolean. By default, a Breaker has no TripFunc.
+type TripFunc func(*Breaker) bool
 
-type Breaker interface {
-	Call(func() error) error
-	Fail()
-	Failures() int64
-	Trip()
-	Reset()
-	Break()
-	Ready() bool
-	Tripped() bool
-	Subscribe() <-chan BreakerEvent
-}
+// Breaker is the base of a circuit breaker. It maintains failure and success counters
+// as well as the event subscribers.
+type Breaker struct {
+	// BackOff is the backoff policy that is used when determining if the breaker should
+	// attempt to retry. A breaker created with NewBreaker will use an exponential backoff
+	// policy by default.
+	BackOff backoff.BackOff
 
-// TrippableBreaker is a base for building trippable circuit breakers. It keeps
-// track of the tripped state and runs the OnTrip and OnReset callbacks.
-type TrippableBreaker struct {
-	// ResetTimeout is the minimum amount of time the Breaker will wait
-	// before allowing the function to be called again
-	ResetTimeout time.Duration
+	// ShouldTrip is a TripFunc that determines whether a Fail() call should trip the breaker.
+	// A breaker created with NewBreaker will not have a ShouldTrip by default, and thus will
+	// never automatically trip.
+	ShouldTrip TripFunc
 
+	failures       int64
+	consecFailures int64
+	successes      int64
 	_lastFailure   unsafe.Pointer
 	halfOpens      int64
+	nextBackOff    time.Duration
 	tripped        int32
 	broken         int32
-	failures       int64
 	eventReceivers []chan BreakerEvent
 }
 
-func (cb *TrippableBreaker) sendEvent(event BreakerEvent) {
-	for _, receiver := range cb.eventReceivers {
-		receiver <- event
-	}
+// NewBreaker creates a base breaker with an exponential backoff and no TripFunc
+func NewBreaker() *Breaker {
+	b := backoff.NewExponentialBackOff()
+	b.InitialInterval = defaultInitialBackOffInterval
+	b.Reset()
+	return &Breaker{BackOff: b, nextBackOff: b.NextBackOff()}
 }
 
-// NewResettingBreaker returns a new ResettingBreaker with the given reset timeout
-func NewTrippableBreaker(resetTimeout time.Duration) *TrippableBreaker {
-	return &TrippableBreaker{ResetTimeout: resetTimeout}
+// NewThresholdBreaker creates a Breaker with a TripFunc that trips the breaker whenever
+// the failure count meets the threshold.
+func NewThresholdBreaker(threshold int64) *Breaker {
+	breaker := NewBreaker()
+	breaker.ShouldTrip = func(cb *Breaker) bool {
+		return cb.Failures() == threshold
+	}
+	return breaker
+}
+
+// NewConsecutiveBreaker creates a Breaker with a TripFunc that trips the breaker whenever
+// the consecutive failure count meets the threshold.
+func NewConsecutiveBreaker(threshold int64) *Breaker {
+	breaker := NewBreaker()
+	breaker.ShouldTrip = func(cb *Breaker) bool {
+		return cb.ConsecFailures() == threshold
+	}
+	return breaker
+}
+
+// NewRateBreaker creates a Breaker with a TripFunc that trips the breaker whenever the
+// error rate hits the threshold. The error rate is calculated as such:
+// f = number of failures
+// s = number of successes
+// e = f / (f + s)
+// This breaker will not trip until there have been at least minSamples events.
+func NewRateBreaker(rate float64, minSamples int64) *Breaker {
+	breaker := NewBreaker()
+	breaker.ShouldTrip = func(cb *Breaker) bool {
+		samples := cb.Failures() + cb.Successes()
+		return samples >= minSamples && cb.ErrorRate() >= rate
+	}
+	return breaker
 }
 
 // Subscribe returns a channel of BreakerEvents. Whenever the breaker changes state,
 // the state will be sent over the channel. See BreakerEvent for the types of events.
-func (cb *TrippableBreaker) Subscribe() <-chan BreakerEvent {
+func (cb *Breaker) Subscribe() <-chan BreakerEvent {
 	eventReader := make(chan BreakerEvent)
 	output := make(chan BreakerEvent, 100)
 
@@ -118,7 +157,7 @@ func (cb *TrippableBreaker) Subscribe() <-chan BreakerEvent {
 
 // Trip will trip the circuit breaker. After Trip() is called, Tripped() will
 // return true.
-func (cb *TrippableBreaker) Trip() {
+func (cb *Breaker) Trip() {
 	atomic.StoreInt32(&cb.tripped, 1)
 	now := time.Now()
 	atomic.StorePointer(&cb._lastFailure, unsafe.Pointer(&now))
@@ -127,48 +166,83 @@ func (cb *TrippableBreaker) Trip() {
 
 // Reset will reset the circuit breaker. After Reset() is called, Tripped() will
 // return false.
-func (cb *TrippableBreaker) Reset() {
-	if cb.Tripped() {
-		cb.sendEvent(BreakerReset)
-	}
+func (cb *Breaker) Reset() {
 	atomic.StoreInt32(&cb.broken, 0)
 	atomic.StoreInt32(&cb.tripped, 0)
-	atomic.SwapInt64(&cb.failures, 0)
+	atomic.StoreInt64(&cb.failures, 0)
+	atomic.StoreInt64(&cb.consecFailures, 0)
+	atomic.StoreInt64(&cb.successes, 0)
+	cb.sendEvent(BreakerReset)
 }
 
 // Tripped returns true if the circuit breaker is tripped, false if it is reset.
-func (cb *TrippableBreaker) Tripped() bool {
-	return cb.tripped == 1
+func (cb *Breaker) Tripped() bool {
+	return atomic.LoadInt32(&cb.tripped) == 1
 }
 
 // Break trips the circuit breaker and prevents it from auto resetting. Use this when
 // manual control over the circuit breaker state is needed.
-func (cb *TrippableBreaker) Break() {
+func (cb *Breaker) Break() {
 	atomic.StoreInt32(&cb.broken, 1)
 	cb.Trip()
 }
 
-// Call runs the given function.  No wrapping is performed.
-func (cb *TrippableBreaker) Call(circuit func() error) error {
-	return circuit()
-}
-
 // Failures returns the number of failures for this circuit breaker.
-func (cb *TrippableBreaker) Failures() int64 {
+func (cb *Breaker) Failures() int64 {
 	return atomic.LoadInt64(&cb.failures)
 }
 
-// Fail records the time of a failure
-func (cb *TrippableBreaker) Fail() {
+// ConsecFailures returns the number of consecutive failures that have occured.
+func (cb *Breaker) ConsecFailures() int64 {
+	return atomic.LoadInt64(&cb.consecFailures)
+}
+
+// Successes returns the number of successes for this circuit breaker.
+func (cb *Breaker) Successes() int64 {
+	return atomic.LoadInt64(&cb.successes)
+}
+
+// Fail is used to indicate a failure condition the Breaker should record. It will
+// increment the failure counters and store the time of the last failure. If the
+// breaker has a TripFunc it will be called, tripping the breaker if necessary.
+func (cb *Breaker) Fail() {
+	atomic.AddInt64(&cb.failures, 1)
+	atomic.AddInt64(&cb.consecFailures, 1)
 	now := time.Now()
 	atomic.StorePointer(&cb._lastFailure, unsafe.Pointer(&now))
 	cb.sendEvent(BreakerFail)
+	if cb.ShouldTrip != nil && cb.ShouldTrip(cb) {
+		cb.Trip()
+	}
+}
+
+// Success is used to indicate a success condition the Breaker should record. If
+// the success was triggered by a retry attempt, the breaker will be Reset().
+func (cb *Breaker) Success() {
+	cb.BackOff.Reset()
+	cb.nextBackOff = cb.BackOff.NextBackOff()
+
+	state := cb.state()
+	if state == halfopen {
+		cb.Reset()
+	}
+	atomic.StoreInt64(&cb.consecFailures, 0)
+	atomic.AddInt64(&cb.successes, 1)
+}
+
+// ErrorRate returns the current error rate of the Breaker, expressed as a floating
+// point number (e.g. 0.9 for 90%), since the last time the breaker was Reset.
+func (cb *Breaker) ErrorRate() float64 {
+	failures := float64(cb.Failures())
+	successes := float64(cb.Successes())
+	total := failures + successes
+	return failures / total
 }
 
 // Ready will return true if the circuit breaker is ready to call the function.
 // It will be ready if the breaker is in a reset state, or if it is time to retry
 // the call for auto resetting.
-func (cb *TrippableBreaker) Ready() bool {
+func (cb *Breaker) Ready() bool {
 	state := cb.state()
 	if state == halfopen {
 		atomic.StoreInt64(&cb.halfOpens, 0)
@@ -177,19 +251,60 @@ func (cb *TrippableBreaker) Ready() bool {
 	return state == closed || state == halfopen
 }
 
+// Call wraps a function the Breaker will protect. A failure is recorded
+// whenever the function returns an error. If the called function takes longer
+// than timeout to run, a failure will be recorded.
+func (cb *Breaker) Call(circuit func() error, timeout time.Duration) error {
+	var err error
+	state := cb.state()
+
+	if state == open {
+		return ErrBreakerOpen
+	}
+
+	if timeout == 0 {
+		err = circuit()
+	} else {
+		c := make(chan int, 1)
+		go func() {
+			err = circuit()
+			close(c)
+		}()
+
+		select {
+		case <-c:
+		case <-time.After(timeout):
+			err = ErrBreakerTimeout
+		}
+	}
+
+	if err != nil {
+		if state == halfopen {
+			atomic.StoreInt64(&cb.halfOpens, 0)
+		}
+		cb.Fail()
+		return err
+	}
+
+	cb.Success()
+	return nil
+}
+
 // state returns the state of the TrippableBreaker. The states available are:
 // closed - the circuit is in a reset state and is operational
 // open - the circuit is in a tripped state
 // halfopen - the circuit is in a tripped state but the reset timeout has passed
-func (cb *TrippableBreaker) state() state {
+func (cb *Breaker) state() state {
 	tripped := cb.Tripped()
 	if tripped {
 		if cb.broken == 1 {
 			return open
 		}
+
 		since := time.Since(cb.lastFailure())
-		if since > cb.ResetTimeout {
+		if since > cb.nextBackOff {
 			if atomic.CompareAndSwapInt64(&cb.halfOpens, 0, 1) {
+				cb.nextBackOff = cb.BackOff.NextBackOff()
 				return halfopen
 			}
 			return open
@@ -199,183 +314,13 @@ func (cb *TrippableBreaker) state() state {
 	return closed
 }
 
-func (cb *TrippableBreaker) lastFailure() time.Time {
+func (cb *Breaker) lastFailure() time.Time {
 	ptr := atomic.LoadPointer(&cb._lastFailure)
 	return *(*time.Time)(ptr)
 }
 
-// FrequencyBreaker is a circuit breaker that will only trip if the threshold is met
-// within a certain amount of time.
-type FrequencyBreaker struct {
-	// Duration is the amount of time in which the failure theshold must be met.
-	Duration time.Duration
-
-	// Threshold is the number of failures Breaker will allow before tripping
-	Threshold int64
-
-	_failureTick      unsafe.Pointer
-	failuresSinceTick int64
-	*TrippableBreaker
-}
-
-// NewFrequencyBreaker returns a new FrequencyBreaker with the given duration
-// and failure threshold. If a duration is specified as 0 then no duration will be used and
-// the behavior will be the same as a ThresholdBreaker
-func NewFrequencyBreaker(duration time.Duration, threshold int64) *FrequencyBreaker {
-	return &FrequencyBreaker{duration, threshold, nil, 0, NewTrippableBreaker(time.Millisecond * 500)}
-}
-
-// Fail records a failure. If the failure count meets the threshold within the duration,
-// the circuit breaker will trip. If a BreakerTripped callback is available it will be run.
-func (cb *FrequencyBreaker) Fail() {
-	if cb.Duration > 0 {
-		cb.frequencyFail()
-	}
-
-	cb.TrippableBreaker.Fail()
-	atomic.AddInt64(&cb.failuresSinceTick, 1)
-	failures := atomic.AddInt64(&cb.failures, 1)
-	if failures == cb.Threshold {
-		cb.Trip()
+func (cb *Breaker) sendEvent(event BreakerEvent) {
+	for _, receiver := range cb.eventReceivers {
+		receiver <- event
 	}
 }
-
-// Failures returns the number of failures for this circuit breaker. The failure count
-// for a FrequencyBreaker resets when the duration expires.
-func (cb *FrequencyBreaker) Failures() int64 {
-	if cb.Duration <= 0 {
-		return cb.TrippableBreaker.Failures()
-	}
-
-	if time.Since(cb.failureTick()) > cb.Duration {
-		atomic.StoreInt64(&cb.failuresSinceTick, 0)
-		return 0
-	}
-	return atomic.LoadInt64(&cb.failuresSinceTick)
-}
-
-func (cb *FrequencyBreaker) frequencyFail() {
-	if time.Since(cb.failureTick()) > cb.Duration {
-		now := time.Now()
-		atomic.StorePointer(&cb._failureTick, unsafe.Pointer(&now))
-		atomic.SwapInt64(&cb.failures, 0)
-	}
-}
-
-// Call wraps the function the FrequencyBreaker will protect. A failure is recorded
-// whenever the function returns an error. If the threshold is met within the duration,
-// the FrequencyBreaker will trip.
-func (cb *FrequencyBreaker) Call(circuit func() error) error {
-	state := cb.state()
-
-	if state == open {
-		return ErrBreakerOpen
-	}
-
-	err := circuit()
-
-	if err != nil {
-		if state == halfopen {
-			atomic.StoreInt64(&cb.halfOpens, 0)
-		}
-
-		cb.Fail()
-
-		return err
-	}
-
-	cb.Reset()
-
-	return nil
-}
-
-func (cb *FrequencyBreaker) failureTick() time.Time {
-	if cb._failureTick == nil {
-		now := time.Now()
-		atomic.StorePointer(&cb._failureTick, unsafe.Pointer(&now))
-		return now
-	}
-	ptr := atomic.LoadPointer(&cb._failureTick)
-	return *(*time.Time)(ptr)
-}
-
-// ThresholdBreaker is a circuit breaker that will trip when its failure count
-// passes a given threshold. Clients of ThresholdBreaker can either manually call the
-// Fail function to record a failure, checking the tripped state themselves, or they
-// can use the Call function to wrap the ThresholdBreaker around a function call.
-type ThresholdBreaker struct {
-	*FrequencyBreaker
-}
-
-// NewThresholdBreaker creates a new ThresholdBreaker with the given failure threshold.
-func NewThresholdBreaker(threshold int64) *ThresholdBreaker {
-	return &ThresholdBreaker{NewFrequencyBreaker(0, threshold)}
-}
-
-// TimeoutBreaker is a ThresholdBreaker that will record a failure if the function
-// it is protecting takes too long to run. Clients of Timeout must use the Call function.
-// The Fail function is a noop.
-type TimeoutBreaker struct {
-	// Timeout is the length of time the Breaker will wait for Call() to finish
-	Timeout time.Duration
-	*ThresholdBreaker
-}
-
-// NewTimeoutBreaker returns a new TimeoutBreaker with the given call timeout and failure threshold.
-// If timeout is specified as 0 then no timeout will be used and the behavior will be the
-// same as a ThresholdBreaker
-func NewTimeoutBreaker(timeout time.Duration, threshold int64) *TimeoutBreaker {
-	return &TimeoutBreaker{timeout, NewThresholdBreaker(threshold)}
-}
-
-// Fail is a noop for a TimeoutBreaker. Clients must use Call()
-func (cb *TimeoutBreaker) Fail() {
-}
-
-// Call wraps the function the TimeoutBreaker will protect. A failure is recorded
-// whenever the function returns an error. If the threshold is met, the TimeoutBreaker
-// will trip.
-func (cb *TimeoutBreaker) Call(circuit func() error) error {
-	c := make(chan int, 1)
-	var err error
-
-	if cb.Timeout == 0 {
-		return cb.ThresholdBreaker.Call(circuit)
-	}
-
-	go func() {
-		err = cb.ThresholdBreaker.Call(circuit)
-		close(c)
-	}()
-	select {
-	case <-c:
-		if err != nil && err != ErrBreakerOpen {
-			cb.ThresholdBreaker.Fail()
-		}
-		return err
-	case <-time.After(cb.Timeout):
-		cb.ThresholdBreaker.Fail()
-		return ErrBreakerTimeout
-	}
-}
-
-// NoOp returns a Breaker null object.  It implements the interface with
-// no-ops for every function.
-func NoOp() Breaker {
-	return noop
-}
-
-type noOpBreaker struct{}
-
-func (c *noOpBreaker) Call(f func() error) error {
-	return f()
-}
-
-func (c *noOpBreaker) Fail()                           {}
-func (c *noOpBreaker) Trip()                           {}
-func (c *noOpBreaker) Reset()                          {}
-func (c *noOpBreaker) Break()                          {}
-func (c *noOpBreaker) Failures() int64                 { return 0 }
-func (c *noOpBreaker) Ready() bool                     { return true }
-func (c *noOpBreaker) Tripped() bool                   { return false }
-func (cb *noOpBreaker) Subscribe() <-chan BreakerEvent { return nil }